backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-huggingface
huggingface_hub
python-dotenv
